Revert "[KNI] pfpstatus: fix naming, add tests"
